debugging validation

+++ solaris/nets/datagen.py
@@ -201,8 +201,8 @@ class TorchDataset(Dataset):
         sample = {'image': image, 'mask': mask}
         if self.aug:
             sample = self.aug(**sample)
-        sample['image'] = _check_channel_order(sample['image'], 'torch')
-        sample['mask'] = _check_channel_order(sample['mask'], 'torch')
+        sample['image'] = _check_channel_order(sample['image'], 'torch').astype(np.float32)
+        sample['mask'] = _check_channel_order(sample['mask'], 'torch').astype(np.float32)
         return sample
